Split out assets package into traits package and data package
