+++ test/connectors/SubscriptionConnectorSpec.scala
@@ -16,13 +16,13 @@
 
 package connectors
 
-import builders.TestUserBuilder
 import models.{AgentSubmissionModel, CompanySubmissionModel, SubscriptionReference, UserFactsModel}
 import org.mockito.ArgumentMatchers
 import org.scalatest.mock.MockitoSugar
 import play.api.libs.json.{JsValue, Json}
 import uk.gov.hmrc.play.http.{HeaderCarrier, HttpResponse}
 import config.{AppConfig, WSHttp}
+import data.TestUserBuilder
 import org.mockito.Mockito._
 import play.api.http.Status._
 import uk.gov.hmrc.play.test.{UnitSpec, WithFakeApplication}
