+++ test/controllers/EnterCorrespondenceAddressControllerSpec.scala
@@ -16,11 +16,10 @@
 
 package controllers
 
-import assets.{ControllerTestSpec, MessageLookup}
 import auth.{AuthorisedActions, CgtNROrganisation}
-import builders.TestUserBuilder
 import config.{AppConfig, BusinessCustomerSessionCache, SubscriptionSessionCache}
 import connectors.KeystoreConnector
+import data.{MessageLookup, TestUserBuilder}
 import forms.CorrespondenceAddressForm
 import org.jsoup.Jsoup
 import org.mockito.ArgumentMatchers
@@ -30,6 +29,7 @@ import org.mockito.stubbing.Answer
 import play.api.mvc.{Action, AnyContent, Results}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
+import traits.ControllerTestSpec
 import types.AuthenticatedNROrganisationAction
 import uk.gov.hmrc.play.frontend.auth.AuthContext
