+++ test/controllers/UserDetailsControllerSpec.scala
@@ -16,11 +16,10 @@
 
 package controllers
 
-import assets.{ControllerTestSpec, MessageLookup}
 import auth.{AuthorisedActions, CgtIndividual}
-import builders.TestUserBuilder
+import data.{MessageLookup, TestUserBuilder}
 import forms.UserFactsForm
-import models.{UserFactsModel, SubscriptionReference}
+import models.{SubscriptionReference, UserFactsModel}
 import org.jsoup.Jsoup
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
@@ -30,6 +29,7 @@ import play.api.mvc.{Action, AnyContent, Results}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import services.SubscriptionService
+import traits.ControllerTestSpec
 import types.AuthenticatedIndividualAction
 import uk.gov.hmrc.play.frontend.auth.AuthContext
 import uk.gov.hmrc.play.http.HeaderCarrier
