+++ test/data/MessageLookup.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package assets
+package data
 
 object MessageLookup {
