+++ test/predicates/NonResidentIndividualVisibilityPredicateSpec.scala
@@ -16,10 +16,10 @@
 
 package predicates
 
-import builders.TestUserBuilder
 import common.Constants.AffinityGroup
 import common.Keys
 import config.AppConfig
+import data.TestUserBuilder
 import models.{AuthorisationDataModel, Enrolment, Identifier}
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
